fast_api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pillow_heif==1.3.0
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
